Added Content-type header for 503 response
Instead of having a "503 Service unavailable" response when it is blacklist, it's being raised a "500 Internal Server Error" with "Rack::Lint::LintError: No Content-Type header found" exception.

I had this problem using Sinatra in development environment and running a Rack server (rackup). It happens because rackup servers automatically add Rack::Lint in development mode.

To ensure Rack protocol is being followed, there must be a Content-type header in all responses, except when the status is 1xx, 204 or 304.